v2.12 rc1

+++ scripts/utils.sh
@@ -2,7 +2,6 @@
 
 BASE_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null 2>&1 && pwd)"
 
-# shellcheck source=./const.sh
 . "${BASE_DIR}/const.sh"
 
 function is_confirm() {
@@ -19,24 +18,20 @@ function random_str() {
   if [[ -z ${len} ]]; then
     len=16
   fi
-  command -v dmidecode &>/dev/null
   uuid=None
-  if [[ "$?" == "0" ]]; then
+  if command -v dmidecode &>/dev/null; then
     uuid=$(dmidecode -t 1 | grep UUID | awk '{print $2}' | base64 | head -c ${len})
   fi
-  if [[ "$(echo $uuid | wc -L)" == "${len}" ]]; then
-    echo ${uuid}
+  if [[ "$(echo "$uuid" | wc -L)" == "${len}" ]]; then
+    echo "${uuid}"
   else
-    cat /dev/urandom | tr -dc A-Za-z0-9 | head -c ${len}; echo
+    cat < /dev/urandom | tr -dc A-Za-z0-9 | head -c ${len}; echo
   fi
 }
 
 function has_config() {
   key=$1
-  cwd=$(pwd)
-  grep "^${key}=" "${CONFIG_FILE}" &>/dev/null
-
-  if [[ "$?" == "0" ]]; then
+  if grep "^${key}=" "${CONFIG_FILE}" &>/dev/null; then
     echo "1"
   else
     echo "0"
@@ -44,7 +39,6 @@ function has_config() {
 }
 
 function get_config() {
-  cwd=$(pwd)
   key=$1
   value=$(grep "^${key}=" "${CONFIG_FILE}" | awk -F= '{ print $2 }')
   echo "${value}"
@@ -79,39 +73,50 @@ function test_mysql_connect() {
   password=$4
   db=$5
   command="CREATE TABLE IF NOT EXISTS test(id INT); DROP TABLE test;"
-  docker run -i --rm jumpserver/mysql:5 mysql -h${host} -P${port} -u${user} -p${password} ${db} -e "${command}" 2>/dev/null
+
+  mysql_images=$(get_mysql_images)
+  docker run -i --rm "${mysql_images}" mysql -h"${host}" -P"${port}" -u"${user}" -p"${password}" "${db}" -e "${command}" 2>/dev/null
 }
 
 function test_redis_connect() {
   host=$1
   port=$2
   password=$3
-  password=${password:=''}
-  docker run -i --rm jumpserver/redis:6-alpine redis-cli -h "${host}" -p "${port}" -a "${password}" info | grep "redis_version" >/dev/null
+  docker run -i --rm jumpserver/redis:6-alpine redis-cli -h "${host}" -p "${port}" -a "${password}" info | grep "redis_version" 2>/dev/null
+}
+
+function get_mysql_images() {
+  if [[ "$(uname -m)" == "aarch64" ]]; then
+    mysql_images=jumpserver/mariadb:10
+  else
+    mysql_images=jumpserver/mysql:5
+  fi
+  echo "${mysql_images}"
 }
 
 function get_images() {
   scope="all"
-  if [[ ! -z "$1" ]]; then
+  if [[ -n "$1" ]]; then
     scope="$1"
   fi
+
+  mysql_images=$(get_mysql_images)
+
   images=(
     "jumpserver/redis:6-alpine"
-    "jumpserver/mysql:5"
-    "jumpserver/nginx:alpine2"
-    "jumpserver/luna:${VERSION}"
+    "${mysql_images}"
+    "jumpserver/nginx:${VERSION}"
     "jumpserver/core:${VERSION}"
     "jumpserver/koko:${VERSION}"
     "jumpserver/lion:${VERSION}"
-    "jumpserver/lina:${VERSION}"
   )
   for image in "${images[@]}"; do
     echo "${image}"
   done
   if [[ "${scope}" == "all" ]]; then
-    echo "registry.fit2cloud.com/jumpserver/xpack:${VERSION}"
-    echo "registry.fit2cloud.com/jumpserver/omnidb:${VERSION}"
-    echo "registry.fit2cloud.com/jumpserver/xrdp:${VERSION}"
+    echo "registry.jumpserver.org/jumpserver/xpack:${VERSION}"
+    echo "registry.jumpserver.org/jumpserver/omnidb:${VERSION}"
+    echo "registry.jumpserver.org/jumpserver/xrdp:${VERSION}"
   fi
 }
 
@@ -120,7 +125,7 @@ function read_from_input() {
   msg=$2
   choices=$3
   default=$4
-  if [[ ! -z "${choices}" ]]; then
+  if [[ -n "${choices}" ]]; then
     msg="${msg} (${choices}) "
   fi
   if [[ -z "${default}" ]]; then
@@ -129,11 +134,11 @@ function read_from_input() {
     msg="${msg} ($(gettext 'default') ${default})"
   fi
   echo -n "${msg}: "
-  read input
-  if [[ -z "${input}" && ! -z "${default}" ]]; then
-    export ${var}="${default}"
+  read -r input
+  if [[ -z "${input}" && -n "${default}" ]]; then
+    export "${var}"="${default}"
   else
-    export ${var}="${input}"
+    export "${var}"="${input}"
   fi
 }
 
@@ -195,7 +200,7 @@ function log_error() {
 
 function get_docker_compose_services() {
   ignore_db="$1"
-  services="core koko lion lina luna nginx"
+  services="core koko lion nginx"
   use_task=$(get_config USE_TASK)
   if [[ "${use_task}" != "0" ]]; then
     services+=" celery"
@@ -224,16 +229,21 @@ function get_docker_compose_cmd_line() {
   cmd="docker-compose -f ./compose/docker-compose-app.yml"
   use_ipv6=$(get_config USE_IPV6)
   if [[ "${use_ipv6}" != "1" ]]; then
-    cmd="${cmd} -f compose/docker-compose-network.yml"
+    cmd="${cmd} -f ./compose/docker-compose-network.yml"
   else
-    cmd="${cmd} -f compose/docker-compose-network_ipv6.yml"
+    cmd="${cmd} -f ./compose/docker-compose-network_ipv6.yml"
   fi
   services=$(get_docker_compose_services "$ignore_db")
   if [[ "${services}" =~ celery ]]; then
     cmd="${cmd} -f ./compose/docker-compose-task.yml"
   fi
   if [[ "${services}" =~ mysql ]]; then
-    cmd="${cmd} -f ./compose/docker-compose-mysql.yml -f ./compose/docker-compose-mysql-internal.yml"
+    cmd="${cmd} -f ./compose/docker-compose-mysql-internal.yml"
+    if [[ "$(uname -m)" == "aarch64" ]]; then
+      cmd="${cmd} -f ./compose/docker-compose-mariadb.yml"
+    else
+      cmd="${cmd} -f ./compose/docker-compose-mysql.yml"
+    fi
   fi
   if [[ "${services}" =~ redis ]]; then
     cmd="${cmd} -f ./compose/docker-compose-redis.yml -f ./compose/docker-compose-redis-internal.yml"
@@ -256,26 +266,25 @@ function get_docker_compose_cmd_line() {
 
 function install_required_pkg() {
   required_pkg=$1
-  if command -v dnf > /dev/null; then
+  if command -v dnf >/dev/null; then
     if [ "$required_pkg" == "python" ]; then
       dnf -q -y install python2
-      ln -s /usr/bin/python2 /usr/bin/python
     else
-      dnf -q -y install $required_pkg
+      dnf -q -y install "$required_pkg"
     fi
-  elif command -v yum > /dev/null; then
-    yum -q -y install $required_pkg
-  elif command -v apt > /dev/null; then
-    apt-get -qq -y install $required_pkg
-  elif command -v zypper > /dev/null; then
-    zypper -q -n install $required_pkg
-  elif command -v apk > /dev/null; then
+  elif command -v yum >/dev/null; then
+    yum -q -y install "$required_pkg"
+  elif command -v apt >/dev/null; then
+    apt-get -qq -y install "$required_pkg"
+  elif command -v zypper >/dev/null; then
+    zypper -q -n install "$required_pkg"
+  elif command -v apk >/dev/null; then
     if [ "$required_pkg" == "python" ]; then
       apk add -q python2
     else
-      apk add -q $required_pkg
+      apk add -q "$required_pkg"
     fi
-    command -v gettext > /dev/null || {
+    command -v gettext >/dev/null || {
       apk add -q gettext-dev
     }
   else
@@ -291,33 +300,34 @@ function prepare_online_install_required_pkg() {
 }
 
 function prepare_set_redhat_firewalld() {
-  if [[ -f "/etc/redhat-release" ]]; then
-    firewall-cmd --state > /dev/null 2>&1
-    if [[ "$?" == "0" ]]; then
+  if command -v firewall-cmd >/dev/null; then
+    if firewall-cmd --state >/dev/null 2>&1; then
       docker_subnet=$(get_config DOCKER_SUBNET)
-      if [[ ! "$(firewall-cmd --list-rich-rule | grep ${docker_subnet})" ]]; then
-        firewall-cmd --permanent --zone=public --add-rich-rule="rule family=ipv4 source address=${docker_subnet} accept"
-        firewall-cmd --reload
+      if ! firewall-cmd --list-rich-rule | grep "${docker_subnet}" >/dev/null; then
+        firewall-cmd --permanent --zone=public --add-rich-rule="rule family=ipv4 source address=${docker_subnet} accept" >/dev/null
+        flag=1
       fi
-      if command -v dnf > /dev/null; then
-        if [[ ! "$(firewall-cmd --list-all | grep 'masquerade: yes')" ]]; then
-          firewall-cmd --permanent --add-masquerade
-          firewall-cmd --reload
+      if command -v dnf >/dev/null; then
+        if ! firewall-cmd --list-all | grep 'masquerade: yes' >/dev/null; then
+          firewall-cmd --permanent --add-masquerade >/dev/null
+          flag=1
         fi
+      fi
+      if [[ "$flag" ]]; then
+          firewall-cmd --reload >/dev/null
+          unset flag
       fi
     fi
   fi
 }
 
 function prepare_config() {
-  cwd=$(pwd)
-  cd "${PROJECT_DIR}" || exit
+  cd "${PROJECT_DIR}" || exit 1
 
-  config_dir=$(dirname "${CONFIG_FILE}")
   echo_yellow "1. $(gettext 'Check Configuration File')"
-  echo "$(gettext 'Path to Configuration file'): ${config_dir}"
-  if [[ ! -d ${config_dir} ]]; then
-    mkdir -p ${config_dir}
+  echo "$(gettext 'Path to Configuration file'): ${CONFIG_DIR}"
+  if [[ ! -d ${CONFIG_DIR} ]]; then
+    mkdir -p "${CONFIG_DIR}"
     cp config-example.txt "${CONFIG_FILE}"
   fi
   if [[ ! -f ${CONFIG_FILE} ]]; then
@@ -331,41 +341,48 @@ function prepare_config() {
   if [[ ! -f "./compose/.env" ]]; then
     ln -s "${CONFIG_FILE}" ./compose/.env
   fi
-  configs=("nginx" "core" "koko" "mysql" "redis")
-  for d in "${configs[@]}"; do
-    for f in $(ls ${PROJECT_DIR}/config_init/${d} | grep -v cert); do
-      if [[ ! -f "${CONFIG_DIR}/${d}/${f}" ]]; then
-        \cp -rf "${PROJECT_DIR}/config_init/${d}" "${CONFIG_DIR}"
-      else
-        echo -e "${CONFIG_DIR}/${d}/${f}  [\033[32m √ \033[0m]"
-      fi
-    done
-  done
-  echo_done
 
-  nginx_cert_dir="${config_dir}/nginx/cert"
-  echo_yellow "\n2. $(gettext 'Configure Nginx')"
-  echo "$(gettext 'configuration file'): ${nginx_cert_dir}"
-  # 迁移 nginx 的证书
-  if [[ ! -d ${nginx_cert_dir} ]]; then
-    mkdir -p "${nginx_cert_dir}"
-    \cp -f "${PROJECT_DIR}"/config_init/nginx/cert/* "${nginx_cert_dir}"
-  fi
-
-  for f in $(ls ${PROJECT_DIR}/config_init/nginx/cert); do
-    if [[ ! -f "${nginx_cert_dir}/${f}" ]]; then
-      \cp -f "${PROJECT_DIR}"/config_init/nginx/cert/${f} "${nginx_cert_dir}"
-    else
-      echo -e "${nginx_cert_dir}/${f}  [\033[32m √ \033[0m]"
+  for d in $(ls "${PROJECT_DIR}/config_init"); do
+    if [[ -d "${PROJECT_DIR}/config_init/${d}" ]]; then
+      for f in $(ls "${PROJECT_DIR}/config_init/${d}"); do
+        if [[ -f "${PROJECT_DIR}/config_init/${d}/${f}" ]]; then
+          if [[ ! -f "${CONFIG_DIR}/${d}/${f}" ]]; then
+            \cp -rf "${PROJECT_DIR}/config_init/${d}" "${CONFIG_DIR}"
+          else
+            echo -e "${CONFIG_DIR}/${d}/${f}  [\033[32m √ \033[0m]"
+          fi
+        fi
+      done
     fi
   done
+
+  nginx_cert_dir="${CONFIG_DIR}/nginx/cert"
+  if [[ ! -d ${nginx_cert_dir} ]]; then
+    mkdir -p "${nginx_cert_dir}"
+    \cp -rf "${PROJECT_DIR}/config_init/nginx/cert" "${CONFIG_DIR}/nginx"
+  fi
+
+  for f in $(ls "${PROJECT_DIR}/config_init/nginx/cert"); do
+    if [[ -f "${PROJECT_DIR}/config_init/nginx/cert/${f}" ]]; then
+      if [[ ! -f "${nginx_cert_dir}/${f}" ]]; then
+        \cp -f "${PROJECT_DIR}/config_init/nginx/cert/${f}" "${nginx_cert_dir}"
+      else
+        echo -e "${nginx_cert_dir}/${f}  [\033[32m √ \033[0m]"
+      fi
+    fi
+  done
+  chmod 644 -R "${CONFIG_DIR}"
   echo_done
 
-  backup_dir="${config_dir}/backup"
+  if [[ "$(uname -m)" == "aarch64" ]]; then
+    sed -i "s/# ignore-warnings ARM64-COW-BUG/ignore-warnings ARM64-COW-BUG/g" "${CONFIG_DIR}/redis/redis.conf"
+  fi
+
+  backup_dir="${CONFIG_DIR}/backup"
   mkdir -p "${backup_dir}"
   now=$(date +'%Y-%m-%d_%H-%M-%S')
   backup_config_file="${backup_dir}/config.txt.${now}"
-  echo_yellow "\n3. $(gettext 'Backup Configuration File')"
+  echo_yellow "\n2. $(gettext 'Backup Configuration File')"
   cp "${CONFIG_FILE}" "${backup_config_file}"
   echo "$(gettext 'Back up to') ${backup_config_file}"
   echo_done
@@ -379,7 +396,7 @@ function echo_logo() {
        ██║██║   ██║██╔████╔██║██████╔╝███████╗█████╗  ██████╔╝██║   ██║█████╗  ██████╔╝
   ██   ██║██║   ██║██║╚██╔╝██║██╔═══╝ ╚════██║██╔══╝  ██╔══██╗╚██╗ ██╔╝██╔══╝  ██╔══██╗
   ╚█████╔╝╚██████╔╝██║ ╚═╝ ██║██║     ███████║███████╗██║  ██║ ╚████╔╝ ███████╗██║  ██║
-  ╚════╝  ╚═════╝ ╚═╝     ╚═╝╚═╝     ╚══════╝╚══════╝╚═╝  ╚═╝  ╚═══╝  ╚══════╝╚═╝  ╚═╝
+   ╚════╝  ╚═════╝ ╚═╝     ╚═╝╚═╝     ╚══════╝╚══════╝╚═╝  ╚═╝  ╚═══╝  ╚══════╝╚═╝  ╚═╝
 
 EOF
 
@@ -400,11 +417,32 @@ function image_has_prefix() {
   fi
 }
 
-function perform_db_migrations() {
+function check_container_if_need() {
+  use_external_mysql=$(get_config USE_EXTERNAL_MYSQL)
+  use_ipv6=$(get_config USE_IPV6)
   volume_dir=$(get_config VOLUME_DIR)
+
+  cmd="docker-compose -f ./compose/docker-compose-redis.yml"
+  if [[ "${use_external_mysql}" == "0" ]]; then
+    cmd="${cmd} -f ./compose/docker-compose-init-mysql.yml"
+    if [[ "$(uname -m)" == "aarch64" ]]; then
+      cmd="${cmd} -f ./compose/docker-compose-mariadb.yml"
+    else
+      cmd="${cmd} -f ./compose/docker-compose-mysql.yml"
+    fi
+  fi
+  if [[ "${use_ipv6}" != "1" ]]; then
+    cmd="${cmd} -f compose/docker-compose-network.yml"
+  else
+    cmd="${cmd} -f compose/docker-compose-network_ipv6.yml"
+  fi
+  ${cmd} up -d
+}
+
+function perform_db_migrations() {
   docker run -i --rm --network=jms_net \
     --env-file=/opt/jumpserver/config/config.txt \
-    -v ${volume_dir}/core/data:/opt/jumpserver/data \
+    -v "${volume_dir}/core/data":/opt/jumpserver/data \
     jumpserver/core:"${VERSION}" upgrade_db
 }
 
@@ -413,7 +451,7 @@ function check_ipv6_iptables_if_need() {
   use_ipv6=$(get_config USE_IPV6)
   subnet_ipv6=$(get_config DOCKER_SUBNET_IPV6)
   if [[ "${use_ipv6}" == "1" ]]; then
-    if [[ ! "$(ip6tables -t nat -L | grep "${subnet_ipv6}")" ]]; then
+    if ! ip6tables -t nat -L | grep "${subnet_ipv6}" >/dev/null; then
       ip6tables -t nat -A POSTROUTING -s "${subnet_ipv6}" -j MASQUERADE
     fi
   fi
